requirements-dev.txt
asgiref==3.2.10
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-cors-headers==3.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
django-rest-framework==0.1.0
django-taggit==0.17.1
djangorestframework==3.11.0
easy-thumbnails==2.7
Pillow==7.2.0
psycopg2-binary==2.8.5
pycountry==20.7.3
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
Unidecode==1.1.1
